Renamed debug scene filename
